Refractor of configserver
