+++ config-server/configserver/__init__.py
@@ -1,1 +1,2 @@
-from .server import main, Server
+from .configserver import main, ConfigServer, start_server
+from .auth import test_auth
