+++ config-server/configserver/__main__.py
@@ -1,3 +1,3 @@
-from main import main
+from .configserver import main
 
 main()
