+++ config-server/configserver/errors.py
@@ -1,0 +1,18 @@
+class InvalidCredentialsError(Exception):
+    pass
+
+
+class NotAuthorisedError(Exception):
+    pass
+
+
+class InvalidRouteUUIDError(Exception):
+    pass
+
+
+class InvalidURLError(Exception):
+    pass
+
+
+class InvalidRouteTokenError(Exception):
+    pass
